backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
pillow
google-cloud-storage